// apps/odr-v2/components/signIn/SignIn.types.ts
export interface SignInPropsModel {
  // buttonClickHandler: React.MouseEventHandler<HTMLButtonElement> | undefined
  email: string
  password: string
}
export interface SignUpPropsModel {
  name: string
  email: string
  password: string
  mobileNumber: string
}

export interface profilePageProp {
  name: string
  mobileNumber: string
  email: string
  flatNumber: string
  street: string
  city: string
  zipCode: string
  state: string
  country: string
}